+++ apps/docs/src/content/04-components/data-visualisation/cartesian-chart/examples/area.tsx
@@ -15,19 +15,19 @@ export default () => {
     {
       Datum: "1.9.2025",
       Datenbanken: 10,
-      Webspace: 15,
-      Email: 5,
+      Webspace: 15.4,
+      Email: 5.1,
     },
     {
       Datum: "2.9.2025",
-      Datenbanken: 32,
-      Webspace: 25,
+      Datenbanken: 32.9,
+      Webspace: 25.6,
       Email: 10,
     },
     {
       Datum: "3.9.2025",
       Datenbanken: 40,
-      Webspace: 20,
+      Webspace: 20.2,
       Email: 8,
     },
   ];
@@ -37,14 +37,31 @@ export default () => {
       <Heading>Speicherplatz</Heading>
       <CartesianChart data={data} height="300px">
         <CartesianGrid />
-        <Area dataKey="Datenbanken" />
-        <Area dataKey="Webspace" color="palatinate-blue" />
-        <Area dataKey="Email" color="tangerine" />
+        <Area dataKey="Datenbanken" unit="%" />
+        <Area
+          dataKey="Webspace"
+          color="palatinate-blue"
+          unit="%"
+        />
+        <Area dataKey="Email" color="tangerine" unit="%" />
         <XAxis dataKey="Datum" />
         <YAxis domain={[0, 100]} unit=" %" />
         <ChartLegend />
         <ChartTooltip
-          formatter={(v, k) => `${k}: ${v} %`}
+          formatter={(value, name) =>
+            `${name}: ${Intl.NumberFormat(undefined, {
+              style: "unit",
+              unit: "gigabyte",
+            }).format(
+              typeof value === "number" ? value : 0,
+            )}`
+          }
+          progressBarFormatter={(value) =>
+            Intl.NumberFormat(undefined, {
+              style: "unit",
+              unit: "gigabyte",
+            }).format(typeof value === "number" ? value : 0)
+          }
         />
       </CartesianChart>
     </Section>