+++ packages/components/src/components/List/components/Footer/components/ShowNextBatchButton/ShowNextBatchButton.tsx
@@ -3,11 +3,11 @@ import React from "react";
 import type { ButtonProps } from "@/components/Button";
 import { useList } from "@/components/List/hooks/useList";
 import locales from "../../../../locales/*.locale.json";
-import { useLocalizedStringFormatter } from "react-aria";
+import { useLocalizedStringFormatter } from "@/components/TranslationProvider/useLocalizedStringFormatter";
 import ButtonView from "@/views/ButtonView";
 
 export const ShowNextBatchButton: FC<ButtonProps> = (props) => {
-  const stringFormatter = useLocalizedStringFormatter(locales);
+  const stringFormatter = useLocalizedStringFormatter(locales, "List");
   const list = useList();
   const isLoading = list.loader.useIsLoading();
   const isInitiallyLoading = list.loader.useIsInitiallyLoading();
@@ -27,7 +27,7 @@ export const ShowNextBatchButton: FC<ButtonProps> = (props) => {
       variant="plain"
       size="s"
     >
-      {stringFormatter.format("list.showMore")}
+      {stringFormatter.format("showMore")}
     </ButtonView>
   );
 };