+++ packages/components/src/components/List/components/Header/components/FilterContextMenu/DateRangeFilterPopover.tsx
@@ -3,7 +3,7 @@ import ButtonView from "@/views/ButtonView";
 import headerStyles from "@/components/List/components/Header/Header.module.css";
 import TextView from "@/views/TextView";
 import { IconFilter } from "@/components/Icon/components/icons";
-import { useLocalizedStringFormatter } from "react-aria";
+import { useLocalizedStringFormatter } from "@/components/TranslationProvider/useLocalizedStringFormatter";
 import locales from "../../../../locales/*.locale.json";
 import { Popover, PopoverTrigger } from "@/components/Popover";
 import styles from "./FilterContextMenus.module.scss";
@@ -20,7 +20,7 @@ export const DateRangeFilterPopover: FC<Props> = (props) => {
 
   const { name, property } = filter;
 
-  const stringFormatter = useLocalizedStringFormatter(locales);
+  const stringFormatter = useLocalizedStringFormatter(locales, "List");
 
   const controller = useOverlayController("Popover");
 
@@ -37,7 +37,7 @@ export const DateRangeFilterPopover: FC<Props> = (props) => {
       <Popover
         placement="bottom end"
         isDialogContent
-        aria-label={stringFormatter.format("list.dateRange")}
+        aria-label={stringFormatter.format("dateRange")}
       >
         <RangeCalendarView
           {...filter.dateRangeOptions}