+++ packages/remote-core/src/connection/connectRemoteIframe.ts
@@ -1,24 +1,27 @@
 import {
+  Version,
   type HostExports,
   type HostToRemoteConnection,
   type NavigationState,
   type RemoteExports,
-  Version,
+  type RemoteExtBridgeConnectionApi,
 } from "@/connection/types";
+import { getWithMergedHostConfig } from "@/ext-bridge/getWithMergedHostConfig";
 import { emptyImplementation } from "@/ext-bridge/implementation";
 import { FlowThreadSerialization } from "@/serialization/FlowThreadSerialization";
-import type { ExtBridgeConnectionApi } from "@mittwald/ext-bridge";
+import { type HostConfig } from "@mittwald/flow-core";
 import type { RemoteConnection } from "@mittwald/remote-dom-core/elements";
 import { ThreadIframe } from "@quilted/threads";
 
 interface Options {
   connection: RemoteConnection;
   iframe: HTMLIFrameElement;
+  hostConfig: HostConfig;
   onReady?: (connection: HostToRemoteConnection) => void;
   onLoadingChanged?: (isLoading: boolean) => void;
   onError?: (error: string) => void;
   onNavigationStateChanged?: (state: NavigationState) => void;
-  extBridgeImplementation?: ExtBridgeConnectionApi;
+  extBridgeImplementation?: RemoteExtBridgeConnectionApi;
 }
 
 export const connectRemoteIframe = (opts: Options): HostToRemoteConnection => {
@@ -29,15 +32,21 @@ export const connectRemoteIframe = (opts: Options): HostToRemoteConnection => {
     onLoadingChanged,
     onError,
     onNavigationStateChanged,
-    extBridgeImplementation = emptyImplementation,
+    extBridgeImplementation: extBridgeImplementationProp = emptyImplementation,
+    hostConfig,
   } = opts;
+
+  const extBridgeImplementation = {
+    ...extBridgeImplementationProp,
+    getConfig: getWithMergedHostConfig(extBridgeImplementationProp, hostConfig),
+  };
 
   const result = {
     thread: new ThreadIframe<RemoteExports, HostExports>(iframe, {
       serialization: new FlowThreadSerialization(),
       exports: {
         ...extBridgeImplementation,
-        setIsReady: async (version = 1) => {
+        setIsReady: async (version = Version.v1) => {
           result.version = version;
           onReady?.(result);
         },
@@ -50,9 +59,11 @@ export const connectRemoteIframe = (opts: Options): HostToRemoteConnection => {
         setNavigationState: async (state) => {
           onNavigationStateChanged?.(state);
         },
+        getHostConfig: async () => {
+          return hostConfig;
+        },
       },
     }),
-    version: 0,
     updateHostPathname: (hostPathname?: string) => {
       if (hostPathname === undefined) {
         return;
@@ -62,6 +73,7 @@ export const connectRemoteIframe = (opts: Options): HostToRemoteConnection => {
         result.thread.imports.setPathname(hostPathname);
       }
     },
+    version: 0,
   };
 
   result.thread.imports.render(connection);