+++ packages/remote-elements/src/auto-generated/RemoteChartTooltipElement.ts
@@ -14,6 +14,8 @@ export class RemoteChartTooltipElement extends FlowRemoteElement<RemoteChartTool
       allowEscapeViewBox: {},
       formatter: {},
       headingFormatter: {},
+      progressBarFormatter: {},
+      showProgressBar: {},
       wrapperClassName: {},
     };
   }